+++ src/util/timer.cpp
@@ -31,8 +31,8 @@
 void gettimeofday(struct timeval* t,void* timezone)
 {       struct _timeb timebuffer;
         _ftime( &timebuffer );
-        t->tv_sec=timebuffer.time;
-        t->tv_usec=1000*timebuffer.millitm;
+        t->tv_sec = static_cast<long>(timebuffer.time);
+        t->tv_usec = 1000 * timebuffer.millitm;
 }
 #else
 #include <sys/time.h>
